% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_boundaries.R
\docType{data}
\name{grade_boundaries}
\alias{grade_boundaries}
\title{University of Toronto letter grades and minimum number grades to achieve them}
\format{A data frame with 13 rows and 3 columns

\describe{
\item{letter_grade}{text}
\item{number_grade}{Minimum number grade required to obtain that letter grade, numeric}
\item{grade_points}{Contribution to grade point average, numeric}
}}
\source{
\url{https://www.utsc.utoronto.ca/registrar/u-t-grading-scheme}
}
\usage{
grade_boundaries
}
\description{
University of Toronto letter grades and minimum number grades to achieve them
}
\keyword{datasets}
